% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReducedHeatmap.R
\name{ReducedHeatmap}
\alias{ReducedHeatmap}
\alias{ReducedHeatmap.Seurat}
\title{Compute Heatmap using dimensional reduction}
\usage{
ReducedHeatmap(object, ...)

\method{ReducedHeatmap}{Seurat}(object, assay = "RNA", slot = "data",
  genes_of_interest = NULL, ident = NULL, reduction = "tsne",
  dimension = 1, ...)
}
\arguments{
\item{object}{Object to visualize}

\item{...}{Additional parameters to pass}

\item{assay}{Assay to use. Default: "RNA"}

\item{slot}{Slot to use. Default: "data"}

\item{genes_of_interest}{genes_of_interest Genes of Interest}

\item{ident}{Object variable to use when determining cell_labels}

\item{reduction}{Reduction to visualize. Default: 'tsne'}

\item{dimension}{The dimension to visualize. Default: 1}

\item{expression_matrix}{Full expression matrix genes (rows) vs. cells (columns). Rownames should be gene names}

\item{dim_embedding}{Cell embedding within the dimension of interest}

\item{cell_labels}{Labels for each cell. While this is generally the cluster identity for a cell, it can be any discrete metadata variable}

\item{enrich}{For every gene in genes of interest, find enrich number of genes that are close to that gene in the distance induced by the 1D t-SNE}

\item{breaks}{Number of bins}

\item{slope}{For better visualization, transform the values with a logistic function. This is the slope of that function.}

\item{intercept}{For better visualization, transform the values with a logistic function. This is the intercept of that function.}

\item{object}{}
}
\description{
Experimental version of t-SNE heatmaps codes
}
